Move Travel Estimates Models
